feat: counts reps in exerciselog
